# openapi.yaml
openapi: 3.1.0
info:
  title: Chatrooms API
  description: The API for Chatrooms.
  version: 1.0.0
servers:
  - url: 'https://api.chatrooms.udupa.dev'
paths:
...
  /rooms:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - Rooms
      summary: Get a summary of all rooms.
      description: Get a summary of all rooms available to the currently logged in user.
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: "#/components/schemas/Room"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      security:
        - cookieAuth: [ ]
      tags:
        - Rooms
      summary: Create a new room.
      description: Create a new room as the currently logged in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomCreate"
      responses:
        "201":
          description: Successfully created room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartialRoom"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid fields provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rooms/{roomId}:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - Rooms
      summary: Get information on a specific room.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to get the information of.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      security:
        - cookieAuth: [ ]
      tags:
        - Rooms
      summary: Update an existing room.
      description: Update an existing room as the currently logged in user.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to update.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomPatch"
      responses:
        "204":
          description: Successfully updated room.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to edit this room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid fields provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      security:
        - cookieAuth: [ ]
      tags:
        - Rooms
      summary: Delete a room.
      description: Delete the room with the specified ID.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to delete.
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successfully deleted room.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to delete this room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rooms/{roomId}/members/:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - Room Members
      summary: Get the members in a specific room.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to get the members of.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rooms/{roomId}/members/{memberId}:
    delete:
      security:
        - cookieAuth: [ ]
      tags:
        - Room Members
      summary: Remove a member from a room.
      description: Remove a member from the room with the specified ID.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to remove a member from.
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          description: The user ID of the member to remove.
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Successfully removed member.
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to remove this member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room or member with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rooms/{roomId}/invites:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - Room Invites
      summary: Get the current invites for this room.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to get the invites for.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  invites:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invite"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to view the invites for this room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      security:
        - cookieAuth: [ ]
      tags:
        - Room Invites
      summary: Create a new invite for this room.
      description: Create a new invite for a room as the currently logged in user.
      parameters:
        - name: roomId
          in: path
          description: The ID of the room to create an invite for.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteCreate"
      responses:
        "201":
          description: Successfully created room invite.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartialInvite"
        "401":
          description: Not Authorised.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User does not have the permission to create an invite for this room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Invalid fields provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: betterauth.session_token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [ string, "null" ]
      required:
        - id
        - username
        - displayName
        - avatarUrl
      example:
        id: "12345676543245"
        username: horizell
        displayName: Horizell
    NullableUser:
      type: [ object, "null" ]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: [ string, "null" ]
      required:
        - id
        - username
        - displayName
        - avatarUrl
      example:
        id: "12345676543245"
        username: horizell
        displayName: Horizell
    RoomCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: [ string, "null" ]
      required:
        - name
      example:
        name: General
        description: A general place for conversation or smth :)
    RoomPatch:
      type: object
      properties:
        name:
          type: string
        description:
          type: [ string, "null" ]
      example:
        description: A general place for conversation or smth :)
    Room:
      type: object
      properties:
        id:
          type: string
        creator:
          $ref: "#/components/schemas/NullableUser"
        name:
          type: string
        description:
          type: [ string, "null" ]
      required:
        - id
        - creator
        - name
        - description
    PartialRoom:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: [ string, "null" ]
        name:
          type: string
        description:
          type: [ string, "null" ]
      required:
        - id
        - creatorId
        - name
        - description
      example:
        id: "12384829543245"
        creatorId: "34734829543245"
        name: General
        description: A general place for conversation or smth :)
    InviteCreate:
      type: object
      properties:
        maxUses:
          type: [ integer, "null" ]
          format: int32
        expires:
          type: [ string, "null" ]
          format: date-time
      required:
        - maxUses
        - expires
      example:
        maxUses: 5
        expires: 2026-02-02T22:00:00Z
    Invite:
      type: object
      properties:
        code:
          type: string
        room:
          $ref: "#/components/schemas/PartialRoom"
        creator:
          $ref: "#/components/schemas/NullableUser"
        uses:
          type: integer
          format: int32
        maxUses:
          type: [ integer, "null" ]
          format: int32
        expires:
          type: [ string, "null" ]
          format: date-time
      required:
        - id
        - code
        - room
        - creator
        - uses
        - maxUses
        - expires
    PartialInvite:
      type: object
      properties:
        code:
          type: string
        roomId:
          type: string
        creatorId:
          type: string
        uses:
          type: integer
          format: int32
        maxUses:
          type: [ integer, "null" ]
          format: int32
        expires:
          type: [ string, "null" ]
          format: date-time
      required:
        - id
        - code
        - roomId
        - creatorId
        - uses
        - maxUses
        - expires
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: i32
        message:
          type: string
      example:
        code: 404
        message: User not found.
...